fix: delete the release in ModelBroadcast
which is not correct semantic of release
